Move buyersguide test mixin
